+++ exchanges_wrapper/__init__.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 """
 REST API and WebSocket asyncio wrapper with grpc powered multiplexer server
- for crypto exchanges (Binance, Bitfinex, OKX, Huobi)
+ for crypto exchanges (Binance, Bitfinex, OKX, Huobi, ByBit)
  Utilizes one connection for many accounts and trading pairs.
  For SPOT market only
 """
@@ -12,7 +12,7 @@ __maintainer__ = "Jerry Fedorenko"
 __contact__ = "https://github.com/DogsTailFarmer"
 __email__ = "jerry.fedorenko@yahoo.com"
 __credits__ = ["https://github.com/DanyaSWorlD"]
-__version__ = "2.1.21rc7"
+__version__ = "2.1.21rc8"
 
 from pathlib import Path
 import shutil